add assertions

+++ src/main/java/duchess/logic/commands/DoneGradeCommand.java
@@ -14,8 +14,8 @@ import java.util.Optional;
  * Marks are given to be added to the grade.
  */
 public class DoneGradeCommand extends Command {
-    private int marksObtained;
-    private int maxMarks;
+    private double marksObtained;
+    private double maxMarks;
     private String moduleCode;
     private int gradeNo;
     private static final String VALID_NUMBER_MSG = "Please supply a valid number.";
@@ -30,7 +30,7 @@ public class DoneGradeCommand extends Command {
      * @param marksObtained marks obtained
      * @param maxMarks      maximum marks obtainable
      */
-    public DoneGradeCommand(String moduleCode, int gradeNo, int marksObtained, int maxMarks) {
+    public DoneGradeCommand(String moduleCode, int gradeNo, double marksObtained, double maxMarks) {
         this.gradeNo = gradeNo - 1;
         this.moduleCode = moduleCode;
         this.marksObtained = marksObtained;
